volcano_plot.R: Make labelling thresholds independent
Previously the --labels option added gene names for genes
over a log2fc threshold and below a p value threshold
Now if --labels is set and neither of --log2fc_threshold or
--pval_threshold are set it uses log2fc > 2 and pval < 1e-5.
However, the thresholds can now be set separately or together.
If both thresholds are set a gene must pass both thresholds to be
labelled.