Create python module for Figshare API
figshare.py contains methods for interacting with the API

upload_files_figshare.py is for uploading files either to a new
entry or to an already existing one.

Figshare-categories.xlsx contains ids for available figshare categories

new_article-template.json can be used to fill in article data for
creating a new entry.