Change how the plots are produced
The script used to create one data frame with all the required
genes and sample info and then subset it for each plot. To reduce
memory usage it now has a function to produce the plot data for a
gene and a function to output a plot for that data frame.
The functions are run over each region using `purrr::walk`